🐛 [iOS] Define method to unwrap media subtype

+++ ios/Classes/core/PMAssetPathEntity.h
@@ -36,7 +36,7 @@
 @property(nonatomic, assign) double lat;
 @property(nonatomic, assign) double lng;
 @property(nonatomic, copy) NSString *title;
-@property(nonatomic, assign) NSUInteger subtype;
+@property(nonatomic, assign) int subtype;
 @property(nonatomic, assign) BOOL favorite;
 @property(nonatomic, assign) BOOL isLocallyAvailable;
